Try to allow only one notification chain per alert
Acknowledge sets the Acknowledged bit if it wasn't and closes the ack
chan. The ack chan is a chan that is kept open for all the
notifications for this alert state. When closed, the notifications will
all shut down.

This should also prevent a bug where a notification goes from crit to
warn to crit without a notification timer triggering. Previously, the
notification would continue on from its current location in the chain
even though it should have started its cycle over again. Now, we close
all notifications as soon as we can so we can begin another chain with
the new crit.