Add warn notifications + refactor
The previous implementation relied on many goroutines communicating via
many channels. During restores and clears these would all have to be shut
down and restarted correctly. Although things appeared to work, I could not
convince myself that it was race or bug free.

This design instead maintains a list of notifications, and intelligently
checks them (without polling) when needed. This list is the same data
structure that we serialize to the state file, so many things become
simpler. Instead of having to worry about each notification's goroutines
and channels, just set the map to what it needs to be and it will notify
correctly. This is easier to reason about and explain, so will hopefully be
easier to work with and fix.