Add a scalar type to support correct types
This is a refactor of the type system. Previously, Numbers were accepted as
arguments to some functions (forecastlr, band, percentile). But Numbers are
results from series reduction functions, and are thus grouped. We got
around this by looking for a NumberNode, which was cheating (but I didn't
realize it).

For example, you could not do this: forecastlr(q("q", "1m"), 1+1) since 1+1
is a binary expression, not a number node. Instead, treat constants that
appear in expressions as scalars. These are only converted to Numbers when
combined with grouped results.

Change the ungroup function to return a scalar, so its output is valid
input to these above functions.