diff returns Y(Last)-Y(First) adjusted to query time.
This is useful for instances where you want change over the period of time. It returns the chronologically last value minus the first value.

Because opentsdb might return a different period of time than requested, diff interpolates the value to be that of the time requested. For example if you request 60 seconds, but the query retruns 30 seconds of data, the result will be multiplied by 2