Use tsTime on the silence page
So we can get the cities list, we need to fetch the schedule in the root
scope. We'll need this soon anyway, to show which silence rules are
currently triggering.

I think there's a race condition since the root scope is not guaranteed to
return its schedule data before the silence data is returned, and thus the
timeanddate var won't be set, but it's working so I'll fix that later.