Swap union subgroup results
Previously, if two results had a similar group (one was a subset of the
other), we would take the smaller of the groups (i.e., the subset). This
produces results with multiple, equivalent groups. Instead, take the
larger of the two sets. This appears to be more correct, but I'm not
convinced that it is better in every case.