cmd/scollector: Remove the lock from sendBatch; log more
Before sendBatch would be guaranteed to not swarm our TSAF server with
many requests because it used the queue lock. Instead, let's just not
spawn it as a goroutine and leave the lock available for the methods
that actually modify the queue. This might help fixing the scollector
not sending bug we saw on the search servers.