cmd/scollector: Attempt to prevent WMI stalls from hanging forever
If WMI is temporarily down, timeout after 20s and allow other things to
try. This should prevent the WMI bug in #11.

We have to do more intelligent handling of the process and readers and
writers. I'm sure there's a race bug in here, but I have fixed all the
crashes I could create.