Correctly set time when changing "now"
This bug manifested when comparing the rule page and expression page.
The output differed. After investigation, it was discovered to be
caused by the DateCache incorrectly munging dates during band
functions. The intent of the DateCache was to change "now" by moving
all the dates back in time. But all it did was hard code the "end" time
to a certain instant. This made the multiple queries from band
identical.

Instead, we now provide a function to set absolute times on a request.
This converts all "XX-ago" and req.End == nil to times. This gives a
few benefits. MiniProfiler results are now correct. Previously they
were lying to us since the request was modified after MiniProfiler
recorded it. Expressions and graphs can still contain relative
discriptions, but the queries they produce are now stamped in time. We
can now easily save exact queries instead of having to recreate them in
time.

fixes #74