Lighter processing of search points
Process one batch at once instead of a point at once. This allows for the
locks to be requested much more infrequently. Also, if we're already
processing some data points, drop the current set. This is fine because
this data comes in frequently, so over time all points will eventually be
added to the set.

The mass goroutine creation was happening during the lock in the Process
function. These changes are designed to prevent that from happening. I am
not sure exactly why that was happening. Current guess is computers are
hard and the lock got into a bad state somehow due to it being access >6k
times per second.