Started work on a view for Alert History.
Since things stay on the dashboard. We will want to be able to see the history of what state the alert has been in. In particular for alerts that don't directly link a timeseries that can be graphed.

A fun aspect I might work on is to have a timeline at the top....