cmd/scollector: Revert to simple os.cpu windows handling
The previous version was incorrect because it did not take into account
timing. However, even if it did it would still be wrong. Between the
time that two points were sent, the counter is still going up. Thus,
the next time it *does* get sent the diff in bosun will be large. Not
sure how to fix.