Process search metrics synchronously
Previously, every call to search.NewSearch (as performed by schedule.Init)
would spawn a new go routine. When that search was no longer used, as in
the case of a rule test which creates a new schedule, the go routine was
not stopped, thus leaking. The purpose of the go routine was to prevent
memory and go routine use when the lock was not being unlocked. Since the
lock bugs are now fixed, we no longer need to do that and can return to
realtime processing.

This does mean that, given a very large load, the search queue may not be
able to keep up. But I would guess at that point many other things (like
OpenTSDB) wouldn't be keeping up, too.