Mark unknown errors as inactive (normal)
Error states are sometimes created with empty tagsets when none can be
determined. We can automatically resolve these by marking them normal when
they would be unknown. This is because it must mean the error has been
resolved, we are getting tagsets back, and thus don't touch these again.

It is possible I have not considered this fully, and there are other
times a unknown error would be correct. I cannot come up with them,
however, so am moving forward.

see #208