Convert date representations to strings
Dates were stored as Date objects. When rendered by the machine's locale,
which for some of us was west of UTC, the date shown in the box would be
one previous to that selected. The queried and in-memory date was correct,
however. Now the displayed date is correct, too.

fixes #225