When loading an alert, optionally load the associated template
When you select an alert, a confirmation comes up asking if you would like to load the template associate with the alert. The confirmation is to ensure that a template you are working on does not get written over. This is a one way operation (alert -> template) and not (template -> alert) since multiple alerts can reference a single template.