cmd/scollector: Made errcheck happy by explicitely ignoring errors
The errors in question don't have to be handled, but by explicitely
ignoring them we communicate the intent of not needing to handle them.

The changes either add handling for errors or explicitely ignores
an error by assigning it to a blank (_).

Other changes made while checking out unhandled errors:

in icmp.go:c_icmp the code was using a deprecated method of registering
a handler. The new method is by assigning to a struct field.