leverage LRU caching for scheduled checks and expr/rule tests
For graphite and opentsdb. Not yet for logstash/ES

* for executing expressions/rules via the web UI, we use
a cache that we retain during the lifetime of bosun
Matt and I decided not to expire the cache at given points (such as
reloading rule page), we might revisit this later.
the only risk is that if you query your store for data -5m to now
and your store doesn't have the latest points up to date,
and then 5m from now you query -10min to -5m you'll get the same
cached data, including the incomplete last points

* for scheduled alert checking, the cache lifetime is only each
specific run so that different queries of the same
data for different templates are optimized and values reused in
templates are guaranteed to be in sync.
we don't make the lifetime longer to prevent affecting
future queries (perhaps a band() a week
from now which queries now-current data we just queried and might be
incomplete)
also there's likely less benefit (low hit rate) in keeping the cache
across alerting runs