cmd/bosun: shorten the action page URLs
This prevents proxies from refusing to relay because the HTTP header is
too large. This is done by reducing the length of the requesting page's
URL. But since bosun communicates data between controllers via the URL, we
now introduce a new key/value store to perform that exchange.

The other implementations for this communication aren't sibling-sibling,
they are parent-child. I didn't find a good way to do sibling-sibling
communication, and this is pretty clear, so I think it's a good choice.

fixes #1053