cmd/scollector: hbase collector was allowing values outside the integer range of opentsdb values
Some hbase region metrics have strange values like 9223372036854776000 which is 193 larger than the max integer value of 9223372036854775807
In Go 1.7 these would be encoded in JSON as 9.223372036854776e+18 but in Go 1.8 they are encoded as 9223372036854776000 which opentsdb
tries to parse as an integer (see http://opentsdb.net/docs/build/html/user_guide/writing.html#integer-values ), which causes datapoints to be rejected with errors such as:

error: queue.go:121: 400 Bad Request
error: queue.go:127: {"error":{"code":400,"message":"One or more data points had errors", ...}

For now this is the only collector we have found having issues so we are going to drop the bad datapoints. In the future we may need to change the custom MarshalJSON code in tsdb.go to make sure values larger than math.MaxInt64 are encoded as floats.