cmd/scollector: haproxy collector compatible with 1.7
The haproxy collector used the numerical positions of the columns
for collecting statistics. This breaks when the order of the columns
changes, which is the case for the CSV format in haproxy 1.7.

The name of the columns provided in the CSV header is now used for
fetching the right statistics.