cmd/bosun, docs: Add execTmpl template function
In this commit I try to partly address an issue raised in the report #2197:
I want to generate a JSON on `actionBodyAck` and send it to an HTTP API.
One of the fields in the JSON payload should contain a message attached
to the performed action and also a list of ids of affected incidents,
e.g.: "User X acknowledged N incidents (ids: 1, 2, 4) with a note: M".
I need to use `makeMap` and `json` functions, which take a "pipeline"
in Go's terms, to build a properly encoded JSON; but I also need a
way to traverse `.States[]` and extract a list of ids in a form of a
"pipeline". I could traverse with a help of the `{{range}}` template
action, but there is no way I can capture the result of that execution.

This patch adds a template function `execTmpl` which executes given Go
text/template provided a data structure and returns a resulting string
which can be kept in a template variable or passed on through the
template pipeline. Another way of looking at it is in-line or
sub-templates.

    {{- $list := execTmpl "{{range .}}<li>{{.Id}}</li>\n{{end}}" .States -}}
    {{- $html := printf "%v<br/>Affected incidents:<ul>%v</ul>" .Message $list -}}
    {{- makeMap "message" $html | json -}}

At first I wanted to take advantage of Go's named sub-templates
(`{{define NAME}}...`) and pass that NAME to `execTmpl` instead of a
literal template. Unfortunately template functions do not get a context
of a template in which they are being invoked, therefore I did not have
an access to the parent template containing sub-templates. I could wrap
it in a closure, but there are too many text/template instances created
in different places in code, so I figured that I will give the current
approach a go first and see how is it accepted.