Fix crashing in Limit and reduce functions
- Limit function can't allow negative value as argument, because
implementation of slices in golang doesn't allow it
- reduce should validate that only seriesValue was passed. Possible case
when we can pass different value (e.g. Numeric) can be because in
state.union (if there is different groups but with equal length) that is
called from walkBinary we can set value of one collection point to NaN
(Numeric). If as example there is Numeric + Series, groups equal 'test'
and 'tets', result can be a.Value == Numeric, b.Value == Numeric. By
condition in walkBinary result will be Numeric. It means that expression
like max(max(graphite(...))/graphite(...)) can pass Numeric to Max
function.