add DelayCloseNormal for alert
set DelayCloseNormal to auto close alerts that have been normal for that duration.

I don't ack or close any incidents manually. When I solve the problem, the incidents go back to normal. All is done. But if I forget to ack/close, the new incident happen and I won't get any notification because  the former is still open.

set this DelayCloseNormal  to `5m` like, it will trigger a delayed close for 5 minutes. If incidents happen in 5 mins, the incident will be reopen and don't send notification. If incidents happened after 5 mins, then it's a new incident and send notification.

Of cause set it to `0s` will be ilegal. Delayed close will have at least one `CheckFrequency` to take effect