cmd/bosun: Show parentheses in exprs
Show parentheses when emitting expressions

Problem
=======

When showing evaluated expressions (e.g. in the computations column when
testing an expression or an alert), the expressions are ambiguous (and
misleading) because they include no parentheses, even where they are
necessary.

E.g. evaluate the following expression:

```
(1 + 2) / 3
```

and the computations column will show:

```
1 + 2 / 3 = 1
1 + 2     = 3
```

Compare to evaluating the expression:

```
1 + (2 / 3)
```

which will show:

```
1 + 2 / 3 = 1.66667
2 / 3     = 0.66667
```

Solution
========

The simplest fix is to wrap all binary expressions in parentheses. This
_does_ emit redundant parentheses, but it's much simpler than figuring
out when precedence rules allow the safe elision of parentheses. So for
example this expression:

```
1 + 2 + 3 + 4
```

will be rendered:

```
(((1 + 2) + 3) + 4)
```

I think any extra verbosity here is worth paying for being explicit and
unambiguous.

An alternative solution would be to apply the parentheses in the
_parent_ node - unconditionally if the parent node is a unary expression
and the child is a binary expression, and conditionally if the parent
node is a binary expression and the child is a binary expression: in
this case parentheses are required on the left child node if it binds
less tightly than the parent's operator, and are required on the right
child node if it binds less or equally tightly than the parent's
operator. This is far too painful to think about so I'm not going to.