Update the documentation for dropbool
Fixes the documentation which described the behaviour of `dropbool`
incorrectly.
`dropbool()` currently drops the values where the value in the second
series set IS zero instead of, as described, dropping those where the
value IS NOT zero. To avoid a breaking change which would give the
function the more intuitive behaviour from its name, this only updates
the documentation.

We should consider fixing this with a breaking change in a future
release.

Fixes #2496